requirements.txt
fastapi
uvicorn
pillow
requests
numpy
python-dotenv